Defactoryfy.
The existing factory/mutator system is cumbersome and doesn't add much.
The original goal was to allow users to enable/disable features, such as
ads and gold, depending on their install needs; instead, we shall move
these features to configuration.

The extra boilerplate needed to build templates, and the fact that the
plugins are often *very* closely tied to the markup of individual pages,
makes the codebase difficult to work with.

In the future, a simple plugin system should be added that allows adding
extra routes, but not necessarily mutating individual components.