Fix double render on initial load
This adds a check to ensure the data actually changes on a page's static `populateData` method.
Previously render was called twice on the initial load.  This was due to
`BaseComponent#shouldComponentUpdate` relying on reference equality, but
`React.Component#setState` creating shallow copies of state.

Unfortunately we were relying on `componentDidUpdate` to fire off `'page:update'`, which only
worked because of the double render.  I've moved the firing of `'page:update'` events to a base
page class inside `componentDidMount` ensuring it is always fires after the first render of a page
component on the client. I really think this event is poorly named and the fact that a dependency
relies on an event to resolve `App#render` is brittle to say the least.

see:

https://github.com/reddit/horse-react/blob/8b53f0d83ce568ebef7a71fd7776866a629543e6/src/client.es6.js#L63