[TEMP] Session conversion does not like * scope
We originally changed from a big list o' scopes to *, because it allows our special client to have all the permissions, including as we add new scopes. This worked fine for the password grant flow, but it was failing on session conversion - whereupon we take a user's reddit session cookie, get a modhash, then get a token with the modhash by posting to `api/v1/authorize` (invalid_scope error.) So, instead, we're going back to a big list o' scopes.

api/v1/authorize should be patched, but this will get things working again, in the meantime.