Move `track` prop to getter on page element
It's both easier to keep track of, and harder to accidentally mess up,
when it's a static property or a getter instead of passed in via props.
Since it's page-specific (used in basepage), it makes sense to be here.
This also fixes an issue where both bodylayout and the page are calling
the tracking pixel.