Make properties on BodyLayout pages update correctly.
`routes.jsx` was setting a key on the BodyLayout it created to wrap
pages. This was to make InfoBar be reused for tracking reasons. This
broke properties set on BodyLayout from updating correctly across pages
(when those properties are promises who’s result depends on the state
of rest-cache). This change makes InfoBar’s tracking independent of
it’s parent and allows promise based props on BodyLayout update
correctly.