Make login/register use real links and redirect back properly.
This changes login/register to be regular links. They no longer require
`data-no-route={ true }` or setting the originalUrl.

Cleans up `app.redirect` in `client.es6.js` and removes duplicated
logic in the click handlers to make this possible.

Adds `registerPath` that automatically is `/register` for apps that use
`/login`.

The app links to `LOGIN_PATH` when there’s a good chance the user is
intending to log in to an existing account. eg if they click ‘log in’
on the subnav or in the user menu. Otherwise when an action that
requires having an account is performed `app.needsToLogInUser`will
redirect to `registerPath`