Feature flagging, phase 1
This adds in feature flagging. This first pass hardcodes configuration
into featureflags.es6.js, but will eventually be moved to an external,
non-source-controlled configuration file.

It also sets up base rules for a list of configured users, whether a
user is logged-in, an employee, admin, or in the beta program, and also
querystring-based flags.