Remove mobile-web-app-capable meta tag to avoid full-screen on iOS
The mobile-web-app-capable tag makes the app open in full-screen
mode when saved to your home screen, but that's not super ideal as
many URLs link out to other sites on reddit, which brings up the
browser chrome anyway, causing an app switch. I think we should
remove this and let it open up in the browser chrome as normal.