Events: add track getter to the pages that did not have it
In basePage the  method is not called unless
there is a  property. This is a hack to make sure
finish is called once since bodyLayout extends basePage.
Since chariot will fix the issue with basePage code running
twice this is an easy way to make sure pageview events fire on
every page.