Add screenview events to popular pages
The pages this adds events to are the following:

- subreddits
- special subreddits (frontpage, all, etc)
- profiles
- messages
- login
- registration
- posting

This list is derived from Justin's comment on this ticket - https://reddit.atlassian.net/browse/DA-638?focusedCommentId=55474&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-55474.

This does things differently than the original events patch. The original
patch had middleware listening for `SET_PAGE`. The problem with this approach
is that pages have very different field requirements. Sometimes the data for
those fields wasn't even available unless it was asynchronously fetched for.
The code was fast becoming a mess of special handling statements.

The approach this takes is to ask for the required data explicitly in the
handler and wait for their responses to make the event request. This obviously
comes at the expense of DRY-ness but I think overall is a win for
clarity/maintainability.

NOTES:

- tracking should only be happening client side. Each tracking call is
  protected by an "is client" check.
- `logClientScreenView` is currently a bit hacky. This is due to an artifact
  of a no-js architecture. Because no-js means that each view has a url, even
  an ephemeral view - like the menu dropdown - has a url, which means their
  handlers get called and the screenview events along with them. Because of
  the ephemeral nature of these views, we really don't want to track them.
  Since we're going to do the clean up for no-js in another pass, we're
  hacking in a solution for now. The code for this hack is designed to be as
  straightforward to remove as possible, which can be seen in the variable
  `tempState` being passed into `logClientScreenView`.
- this also turns on the DNT module ("do not track") since this is data that
  is tracked.