Handle comment threads that are too deep
When a comment thread goes deeper than the maximum
levels to render a tree, it was displaying
"More comments (0)" and the action was failing.

This makes it so the behavior is more in line with
the desktop site in which it will load the comment
tree from the given point after clicking a
"Continue this thread" message.