Add the amphtml link element for self posts
For self post pages rendered on the server, we need to add a link
element pointing to the equivalent AMP page, which is how we make the
AMP page visible to the world. Without a configured value, we will not
include an element, so this can be turned on/off via the AMP environment
variable.