Give a nested superagent its own heads object
It looks like superagent gets mad when an object
is changed and passed to a nested superagent. This
ends up giving an error like

Error: `value` required in setHeader("x-modhash", value).

Instead we make a fresh headers object to pass to
the nested superagent. A side-effect of our code
is that catching such an error ends up restarting
the worker, so this causes needless restarts.