Improve error-logging
The goal of this patch is to catch more errors and log them better.
A few bugfixes where done: client actually logs errors, error-logger
middleware now attches .catch to the right place, fixes to parsing
stack traces, extra info when plain objects are rejected, fixed
the try/catch in the timing middleware.

To get the best possible stack traces we re-throw errors so chrome's
built in console catches and formats it. To get really good stack
traces you should enabled the 'Async' button in the sources panel in chrome.