Revert to `secureProxy` cookies flag.
The `secureProxy` flag is for when you have an insecure connection
between your load balanacer / proxy and app server, but still
want to send secure https cookies. It's deprecated so we'll have
to do some work to make it possible to do the correct override for this
in Koa@2.x, but for now this is the best solution.