Gather and send handler timings to the timings endpoint
We add an optional third argument, an object, to the
routes where we can specify arbitrary route metadata
information. In this case, a "name" key specifies the
name of the route for capturing timing info on the
duration of a handler.

Hook into this new feature of @r/platform, prepare
and validate the data and send it down.