Login - fix validation issue
This was failing because a real error object was expected by the
handling code. To fix it, I throw a ValidationError and check for that
type in the handling code. If it's that type, then handle gracefully.

NOTE: this is fairly hacky for the time being as we need this up asap.