Don't produce a canonical when we're missing data
In the case that we're rendering for SEO and they
were hitting a banned subreddit (or the like), we
were barfing on trying to get the cleanPermalink
(in the comments case).  Instead, when looking an
comments and listings of potentially banned subreddits,
we recognize that the data is missing and just don't
make the canonical links.