Use one loading snoo on comments pages
Problem:
We're displaying two snoos when loading comment pages via the url, one
for the post itself and one for the comments section. This doesn't occur
when we come in via a listing. This seems to be because the post is
already loaded as a prereq for the listing page.

Fix:
Check for the post's existence first. If the post doesn't exist at all -
in the case we come in via the url - the post and the comments will come
in at the same time from our api. In that case show only the loading
icon.  If the post does exist - in the case we come in via the listing -
show the post component and show the loading snoo below, where the
comments would be.