Make sure client can update cookies
All of our cookies were getting set with the `httpOnly` flag.
This was intended to increase security, but had the sideeffect
of preventing the client from updating cookies. A symptom of this
is your theme or compact preference not sticking. Session
cookies are still written with the `httpOnly` flag to protect
your tokens.