Normalize z-indexes across app
This is a prep patch/cleanup for the way we handle z-indexes. The goal
is to have a finite and easy to remember amount of stacking levels
components can live at. This is to lower the chance of unforeseen visual
bugs future devs might run into. The z-indexes are meant to be in a
stacking context that is always relative to the dom root and therefore
easier to reason about.

Based on a 50 point scale - with the majority of components having the
implicit z-index of 0 - the z-indexes of anything above the default are
as follows (the higher in the list, the higher the stacking precedence):

- Toasters: 50
- Modals: 40    (Reporting, CaptchaBox)
- Nav Bar and menus: 30
- Dropdowns/tooltips: 20
- Opaque overlay: 10

These are all defined as variables in variables.less. Also these are
meant to handle the most general cases, which is what our current
situation is now anyway, and not meant as a strict ruleset.