Fix missing "load more" button
Bug:
When a comment had no replies but needed a "load more" button, we were
not showing one. This was due a `replies.length` check that was
happening ahead of rendering the "load more" button.

Fix:
Allow the "load more" button to render independent of the length of
replies. This is done by making the `replies.length` check or'd with a
`loadMoreIds.length` check.