Call UserActivity handler on the client for u/ links
We started redirecting `u/*` links to `user/*` because that's the cannonical
url. The redirect was using SET_PAGE, so on the client handlers weren't being
run. This made the user page sit idlely with a loading indicator forever
when you came there via the PostHeader link or u/ links in UGC text.
As a safety measure this patch also changes the PostHeader links to use
the canonical format.