Comment and post events
Feature:
This adds comment events via the `reply` action and post events via the
`submitPost` action.

Details:
- With comment events, you can have parent data values that are derived
from the OP's data or just a parent comment. There's some logic around
that.
- With post events, some keys change depending on whether a post is a
link or a self, so there's some logic around that as well.
- I noticed I was missing `base_url` in the `getBasePayload` function,
so I added that in.

Testing:
For comments, tested on replies to comments and posts and for posting,
tested on link and self posts.