Port the Native Mobile Banner over
Some changes needed for where things were located,
like imports or certain values.  Added a
getRouteMetaFromState helper to assist with knowing
where we're coming in. Made use of the same animation
bits as the toaster and overall restructured things
to be a part of a redux app instead of a separate
standalone wart on the DOM.