Normalize all input font sizes to prevent zooming
iOS will automatically zoom in when an input is selected but not reset
the zoom level when the user is done. This allows for horizontal panning
which is not expected behavior from a UX perspective. (Also, it annoys
the crap out of me :P)

Rationale:
I'm normalizing all inputs because I'd never expect a dev of any
experience level to know/remember this while writing an input
component. By doing this everywhere, it's just taken care of. And b/c
it's css, it's easy to overwrite should the need arise.

It's pretty annoying that iOS forces this but it still looks pretty good
so I feel more ok making a universal rule.

Details:
- This applies to text, number, and password inputs and also textareas.
I specified types on the inputs to limit the scope of change (there's a
ton of input types that I don't want to mess with).