Increase toaster font size
Seeing the toaster pop up on my iPhone 6, I felt like I was squinting
to read the text. Increasing the font-size up to 15px puts it at a
comfortable reading level. This is especially important with the
toaster because it auto-dismisses after 3 seconds.

I tested this down to iPhone 4 dimensions with some of the longer
messages. Sometimes the toaster text ended up taking 3 lines but still
looked pretty good even in those cases.

I removed the right padding on the text because there's already padding
around the close button to right, which was making it look like there
was too much padding in total between the two elements.