Add search events
The search event types are:
- `cs.search_executed`
- `cs.search_cancelled`
- `cs.search_opened`

Caveats:
In order to track search bar closing accurately, I needed a specific
"close search" action to fire. What was available to me was a more
generic action for closing the overlays. To get around this, I pass in
the "close search" handler from the SearchBarOverlay into the generic
OverlayMenu component. If the close handler from the outside props
exists then it uses that else it uses the generic one.