Small tweak to the formatDifference library
The function signature for Date.setMinutes is
Date.setMinutes(min,sec,millisec), and
Date.getTimezoneOffset returns the value in minutes.

Therefore in California, being 7 hours off UTC at
the moment, the timezoneOffset is 420 minutes...
but if you put that as seconds, it ends up being
7 minutes.  Because of that, times being returned
by this library were off by 7 minutes.  Removing
the errant line fixes the problem.