Guard postMessage'ing gtm jail if it doesn't have contentWindow
GTM is proxies its event tracking through an iframe via postMessage, seeing some errors in the logs where gtmIFrame.contentWindow is undefined. I think this is either because the frame times out or the browser isn't letting us access the iframe. The latter might require more work so this is intended to be a stopgap for now so the logs are cleaner