Preserve dissapearing post selftext and previews
This patch solves two frequent instances of self text
and previews dissapearing. The first case is voting
on a selfpost often makes selftext disappear. This seems
like partially a bug from api-client, the version is bumped herein.
The other case is some api endpoints don't return the full media
preview. This mostly seems to impact NSFW posts returned by
the comments page api. You could see this when a NSFW post
had a preview in the listing, but lost it when going to the comments page.