GTM: Set frame name properly
The GTM iframe's `name` property includes a base payload for all
events fired by GTM. Having this set improperly means that
`subreddit`, `origin`, and `pathname` were null for all GTM macros.

This was most noticable for comScore as it requires the `origin`
to detect which site the pixel is being fired from since they all
come from an iframe hosted on redditmedia.com.