Add back the desktop redirect on mdot servers
This fixes a regression from 1X. When a user arrives via an m link url
and they have the mweb-no-redirect cookie, the most expected behavior
is to redirect to desktop, their preferred experience. We're currently
not respecting that preference.

Details:
This adds the redirect to `preRouteServerMiddleware`, which is where
we're doing all our global route handling. Just by virtue of the cookie
existing the redirect should happen. If it doesn't then continue on with
the mweb experience.