Speed up voting
Problem:
Voting is noticeably slow sometimes and very bad on older phones. This
is due to some sloppy selectors in our CommentsPage code - namely the
features selector and the comments selector, but in general most of the
selectors weren't being selective enough.

The features selector was listening for ANY change in state. A simple
upvote fits this criteria and therefore the whole page was re-rendering.
Listening for comments has a similar problem since a new comment, which
is created when we vote, means the whole tree changes.

Fix:
Remove these two selectors and make the rest of the selectors more
specific. This clean up should have the effect of speeding up any
actions done on the comments pages.

Additional changes
I've also cleaned up the voting component since its using the no-js
way of doing things was also a source of slowness. I've simplified it
down to the bare essentials (the amount of code to get this to work
before was ridiculous) and per Erica's recommendation, posts and
comments have the same CSS.