Update page titles to match desktop's
Right now, mweb titles are different than desktop's. For instance, on
the frontpage, currently we have "Reddit - the Front Page of the
Internet", but desktop has "reddit - the front page of the internet".
This updates that title accordingly.

On comment's pages, I removed the "reddit" prefix entirely and moved
the post title to the first position to differentiate tabs better. This
also matches desktop's format.