Session refreshing on visibility changes
Bug:
When a user has a reddit tab open and comes back to it some time later,
the session has expired. Any logged in action that user tries kicks them
to the register screen because our js doesn't know to try to update the
session. A user hitting refresh fixes that since the cookies are still
valid.

Fix:
This adds a visibility check that fires every time a tab is focused,
leveraging the [visibility
api](https://developer.mozilla.org/en-US/docs/Web/API/Page_Visibility_API),
which is [well supported](http://caniuse.com/#search=visibilitychange)
on mobile If the user had a refresh token and the token was expired, the
app will fetch a new one upon being focused again. This is a nice all
around solution since it avoids having to specially handle every logged
in action.

Caveats:
I've cleaned up the session refresh code a bit, which was handling token
refreshing for new and very long sessions. One thing I added was a
minimum to the refresh timer just for fear of ddos'ing ourselves. This
is set at 5 minutes.

Testing:
I tested this on my iPhone by dropping the expiration down to 10
seconds, opening the app in one tab, going to another, then coming back
to the first after 10 seconds. Then hitting upvote. I did a similar test
with the open app and the lock screen. Both worked. I did this in both
Chrome and Safari. This still needs testing on android.