Fix message thread replying
It turns out that we can only reply to a message
thread if the other side has sent us a message.
For example, if there's one message in the thread
and the current user sent it, they can't send
further messages until the other side sends a reply.

Take a look at all of the messages in a thread, find
the latest response from the other side and use that
as the message id we respond to. Unfortunately the
API appears to succeed if we reply to our own message,
so this circumvents the issue.

We no longer show the replyToggle if the user can't
yet reply to a thread.