Unified urls - Match subreddit sorts to desktop
Bug:
Currently 2X doesn't support urls that have the format of `/r/nba/top`,
where the subreddit sort is part of the path. As part of the unified url
push, we need to support this.

Fix:
Currently we support the sort as a query parameter but we'd like to
deprecate this over time. So we're going to first look for the sort in
the url then fallback to the query param then fallback to 'hot'.

There's a few other sorts that 2X doesn't support yet at all, such as
"gilded" or "ads". Per discussion with William, we're going to 404 these
(this is the status quo).