Restrict the amphtml link tag to 2% of content
Create a basic page-bucketing mechanism, so we can restrict a feature to
a percentage of content. This is a temporary mechanism and not a
full-fledged page-based experimentation/flagging framework.

Restrict the amphtml link tag to just 2% of content, so that not all of
our self posts show up in search results as their AMP'd versions.