Fix tracking bug for outbound links with nested content
Outbound links with nested content currently don't work.
This wasn't noticed early enough because this only
occurs in linkbars on the comments page, but this
fix future proofs us for more complicated dom trees.
Nested elements were the target of interaction events,
leading to the code wrongly setting the href of `<span />`s
and etc as opposed to the expected `<a />` tag.