Turn select xpromo variants to 100%
As a temporary measure, we are showing the list/subreddit treatments
(frontpage/listings experiments) also to users bucketed into the control
groups. We want to continue to get bucketing events, so we know when a
user has been exposed to this feature, and we want to show 100% of users
the select treatments. We can't eliminate control groups in the API's
bucketing mechanism, so we use this workaround instead.