Support xpromo interstitials for in-app navigation
We need to support display of the cross-promo interstitials for the
appropriate pages for in-app navigation, but we also don't want to show
multiple interstitials per session (and in particular not for sequential
navigations). We track if we have already shown an xpromo experience,
and then we can avoid showing it for later pages.