Try to replace Reddit-hosted GIFs with MP4s during preview playback (#931)
PostContent tries to replace Reddit-hosted GIFs with MP4s

This can save on bandwidth significantly, due to MP4s's much more efficient compression (compared to GIFs at least). We may show users smaller or lower quality content as a result, but the vast savings on bandwidth and load times present a huge upside which will be helpful to users.