Stop using node's Buffer and crypto modules for hashing events.
This was discovered using webpack's bundle analyzer. There were large noticeable
chunks of elliptic hash functions and bn.js (bignum). We have a Base64 lib
already, so we can stop using Buffer for base64 conversion. To remove crypto,
we can pull in a pure js implementation. This lets only pull in the sha256,
sha1, and HmacSha1 functions we need.

Before and After:
* 357 KB gzipped -> 264 KB gzipped
* 1.3 MB -> 1017 KB