Inline @r/api-client
Reduce the size of api-client, and lodash imports by inlining it.
This also has a benefit for our workflow as we no longer have to
go through the cycle of update npm module -> update this repo ->
update deps repo to deploy changes and fix bugs.

With those changes, the payload size decreased as follows:
  * 1017KB         -> 985KB
  * 264KB gzipped  -> 255KB gzipped