Gracefully handle failed accounts and invalid loidCreated.
There's a fairly frequent 'invalid time' exception being thrown in
`dispatchInitialUser`, that has the effect of returning no content
and issuing 404s. This seems to be often be associated with a
`FAILED_ACCOUNT` action on the redux stack for Googlebot.
So it's possible they're coming in with an invalid loidcreated cookie
that's causing `Date` to throw.

This patch:
* tracks account request failure status, and doesn't set loid
cookies when there was a failure.
* doesn't set loid cookies if the `me` account request failed
* adds try/catch w/logging when doing loidcreated conversion
* fixes a bug in reducers/loid that was using 'loid' and 'loidcreated'
from any recevied account (it should only be checking accounts.me)