Fix toggling subscription
makeUUID() can be called from a few places. In some cases,
we have the server response, and in others an already
cleaned object.  Unfortunately, there's a difference
between "display_name" and "displayName", which was
causing us to set the model's uuid to `undefined`
whenever we'd toggle the subscription.  Instead, if
we already have a uuid, just use that.