Add npm scripts to help lint changed files
We have a lot of linter warnings. We should likely clean many of those
up, but in the meantime the warnings make it difficult to notice
linter warnings in files that you are actually changing. These two
scripts will let developers lint the unstaged files or the set of files
that differ from the main branch. We can eventually incorporate these
checks into a git pre-commit hook or similar.